Add LatLng conversion method and update dependencies
A LatLng conversion method is added to the Gimcama model, enabling the conversion of x and y coordinate strings to a LatLng object for Google Maps. Also, multiple dependencies are updated in the pubspec.lock file including the addition of the google_maps_flutter package. Furthermore, a new providers.dart file was created for exporting the firebase_provider.dart.